Move shouldDehydrateQuery and friends to its own file
