+++ client/state/query-client.ts
@@ -3,30 +3,7 @@ import { QueryClient } from 'react-query';
 import { persistQueryClient } from 'react-query/persistQueryClient-experimental';
 import { shouldPersist, MAX_AGE, SERIALIZE_THROTTLE } from 'calypso/state/initial-state';
 import { getPersistedStateItem, storePersistedStateItem } from 'calypso/state/persisted-state';
-
-const SHOULD_PERSIST_QUERY_KEY = 'shouldPersistQuery';
-
-type PersistencePredicate< T > = ( data: T ) => boolean;
-
-type PersistenceMetaPartial< T > = {
-	[ SHOULD_PERSIST_QUERY_KEY ]: PersistencePredicate< T > | boolean;
-};
-
-const hasPersistenceSetting = ( value: unknown ): value is boolean => {
-	return typeof value === 'boolean';
-};
-
-const hasPersistenceCustomSetting = (
-	value: unknown
-): value is PersistencePredicate< unknown > => {
-	return typeof value === 'function';
-};
-
-export const shouldPersistQuery = < T >(
-	predicate: PersistencePredicate< T > | boolean
-): PersistenceMetaPartial< T > => ( {
-	[ SHOULD_PERSIST_QUERY_KEY ]: predicate,
-} );
+import { shouldDehydrateQuery, shouldPersistQuery } from './should-dehydrate-query';
 
 export async function createQueryClient( userId: number | undefined ): Promise< QueryClient > {
 	const queryClient = new QueryClient( {
@@ -51,25 +28,11 @@ export async function createQueryClient( userId: number | undefined ): Promise< 
 			persistor,
 			maxAge: MAX_AGE,
 			dehydrateOptions: {
-				shouldDehydrateQuery: ( query ) => {
-					if ( query.state.status !== 'success' ) {
-						return false;
-					}
-
-					const shouldPersist = query.meta?.[ SHOULD_PERSIST_QUERY_KEY ];
-
-					if ( hasPersistenceSetting( shouldPersist ) ) {
-						return shouldPersist;
-					}
-
-					if ( hasPersistenceCustomSetting( shouldPersist ) ) {
-						return shouldPersist( query.state.data );
-					}
-
-					return true;
-				},
+				shouldDehydrateQuery,
 			},
 		} );
 	}
 	return queryClient;
 }
+
+export { shouldPersistQuery };
